Fix openssh-server.
In the latest Ubuntu release, the installation of openssh-server automatically includes tzdata. However, it's important to be aware that tzdata requires user interaction, which can potentially lead to a temporary blockage during the installation process. Fortunately, this issue can be resolved by following the steps outlined in the following thread: https://stackoverflow.com/questions/44331836/apt-get-install-tzdata-noninteractive.

Signed-off-by: siaimes <34199488+siaimes@users.noreply.github.com>